# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TicketMate API",
    "version": "1.0.0",
    "description": "API documentation for TicketMate backend services",
    "contact": {
      "name": "TicketMate Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439012"
          },
          "title": {
            "type": "string",
            "example": "Pretty Girls Love Amapiano"
          },
          "date": {
            "type": "string",
            "example": "June 27th"
          },
          "time": {
            "type": "string",
            "example": "1:15 PM - 4:50 AM"
          },
          "location": {
            "type": "string",
            "example": "Accra"
          },
          "venue": {
            "type": "string",
            "example": "NO.5 Bar And Restaurant"
          },
          "price": {
            "type": "string",
            "example": "GH₵950"
          },
          "rating": {
            "type": "string",
            "example": "4.8"
          },
          "attendees": {
            "type": "string",
            "example": "4000+"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "example": "https://images.unsplash.com/photo-1470229722913-7c0e2dbbafd3"
          },
          "category": {
            "type": "string",
            "enum": [
              "Music",
              "Sports",
              "Arts",
              "Education",
              "Food",
              "Tech"
            ],
            "example": "Music"
          },
          "isPopular": {
            "type": "boolean",
            "example": true
          },
          "attendeeImages": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "example": [
              "https://randomuser.me/api/portraits/men/32.jpg"
            ]
          },
          "tickets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "VIP"
                },
                "price": {
                  "type": "string",
                  "example": "GH₵1500"
                },
                "description": {
                  "type": "string",
                  "example": "VIP access with perks"
                },
                "available": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "description": {
            "type": "string",
            "example": "Pretty Girls Love Amapiano is a music and dance event organised by the party invasion team of Ghana."
          },
          "organizer": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "AfroNation Events"
              },
              "avatar": {
                "type": "string",
                "format": "uri",
                "example": "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e"
              },
              "isVerified": {
                "type": "boolean",
                "example": true
              },
              "followers": {
                "type": "string",
                "example": "2.5K"
              },
              "events": {
                "type": "string",
                "example": "12"
              }
            }
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "latitude": {
                "type": "number",
                "example": 5.6037
              },
              "longitude": {
                "type": "number",
                "example": -0.187
              }
            }
          },
          "createdBy": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "cancelled"
            ],
            "example": "published"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventInput": {
        "type": "object",
        "required": [
          "title",
          "date",
          "time",
          "location",
          "venue",
          "price",
          "image",
          "category",
          "description"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "example": "Amazing Music Festival"
          },
          "date": {
            "type": "string",
            "example": "December 25th"
          },
          "time": {
            "type": "string",
            "example": "7:00 PM - 11:00 PM"
          },
          "location": {
            "type": "string",
            "example": "Accra"
          },
          "venue": {
            "type": "string",
            "example": "National Theatre"
          },
          "price": {
            "type": "string",
            "example": "GH₵500"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/event-image.jpg"
          },
          "category": {
            "type": "string",
            "enum": [
              "Music",
              "Sports",
              "Arts",
              "Education",
              "Food",
              "Tech"
            ],
            "example": "Music"
          },
          "description": {
            "type": "string",
            "example": "A fantastic music event with top artists"
          },
          "isPopular": {
            "type": "boolean",
            "example": false
          },
          "attendeeImages": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "tickets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "General"
                },
                "price": {
                  "type": "string",
                  "example": "GH₵500"
                },
                "description": {
                  "type": "string",
                  "example": "General admission"
                },
                "available": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "organizer": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Event Organizers Ghana"
              },
              "avatar": {
                "type": "string",
                "format": "uri"
              },
              "isVerified": {
                "type": "boolean",
                "example": false
              },
              "followers": {
                "type": "string",
                "example": "1K"
              },
              "events": {
                "type": "string",
                "example": "5"
              }
            }
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "latitude": {
                "type": "number"
              },
              "longitude": {
                "type": "number"
              }
            }
          }
        }
      },
...
      "EventsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "count": {
            "type": "number",
            "example": 10
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          }
        }
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "summary": "Get all events",
        "description": "Retrieve all published events from the database",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new event",
        "description": "Create a new event (requires authentication)",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EventResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Event created successfully"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}